Update file.List to include directories
