+++ file/local/read.go
@@ -6,9 +6,7 @@ import (
 	"crypto/md5"
 	"hash"
 	"io"
-	"io/ioutil"
 	"os"
-	"path"
 	"path/filepath"
 
 	"github.com/pcelvng/task-tools/file/stat"
@@ -127,36 +125,4 @@ func (r *hashReader) Read(p []byte) (n int, err error) {
 	// writing nothing doesn't affect the final sum
 	r.Hshr.Write(p[:n])
 	return
-}
-
-// ListFiles will list all files in the provided pth directory.
-// pth must be a directory.
-//
-// Will not list recursively and does not return directories.
-// Checksums are not returned.
-func ListFiles(pth string) ([]stat.Stats, error) {
-	// remove local:// prefix if exists
-	pth = rmLocalPrefix(pth)
-
-	pth, _ = filepath.Abs(pth)
-	filesInfo, err := ioutil.ReadDir(pth)
-	if err != nil {
-		return nil, err
-	}
-
-	allSts := make([]stat.Stats, 0)
-	for _, fInfo := range filesInfo {
-		if fInfo.IsDir() {
-			continue
-		}
-
-		sts := stat.New()
-		sts.SetCreated(fInfo.ModTime())
-		sts.SetPath(path.Join(pth, fInfo.Name())) // full abs path
-		sts.SetSize(fInfo.Size())
-
-		allSts = append(allSts, sts)
-	}
-
-	return allSts, nil
 }
